+++ src/views/ContextMenu/ContextMenu.hpp
@@ -9,7 +9,7 @@
 #ifndef ddui_ContextMenu_hpp
 #define ddui_ContextMenu_hpp
 
-#include <ddui/ddui>
+#include <ddui/core>
 #include <functional>
 #include <string>
 #include <vector>
@@ -24,7 +24,7 @@ struct Item {
 void update(std::function<void()> inner_update);
 
 int process_action(void* identifier);
-void show(void* identifier, int x, int y, std::vector<Item> items);
+void show(void* identifier, float x, float y, std::vector<Item> items);
 
 }
